backend/requirements.txt
django>=4.2
psycopg2-binary>=2.9
pandas>=2.0
django-tailwind
djangorestframework
djangorestframework-simplejwt
gunicorn
Django>=4.2
django-cors-headers
django-prometheus
celery==5.3.4
redis==5.0.1
django-guardian==2.4.0
django-simple-history
channels
channels_redis
